Initial commit for ESP32 catmini architecture
With this new architecture I pretend to create a project based on
classes. I have selected Eclipse with Sloeber plugin as development
enviroment. I have created a simulated thread structure.

	-	Change list:
	
		*	Folder structure by functionalities
		*	A controller for every basic functionality
				-	Gait controller 
				-	Serial controller
				-	Servo controller
				-	Accelerometer controller
		*	A class bean for every physical or logical object 
				-	Gait
				-	Servo
				-	...
		*	Logger function

This code is not completely functional right now. This only suppose a
solid base base to complete the rest of functionalities.